Rename test utils to helpers
I need to add a utils folder to the main project, so this will avoid the
two conflicting when I come to test a utility class.
