+++ GetIntoTeachingApiTests/Controllers/CandidatesControllerTests.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using Microsoft.AspNetCore.Mvc;
 using FluentAssertions;
-using GetIntoTeachingApiTests.Utils;
+using GetIntoTeachingApiTests.Helpers;
 using GetIntoTeachingApi.Services;
 using System.Collections.Generic;
