+++ GetIntoTeachingApiTests/Services/CrmCacheTests.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using FluentAssertions;
 using GetIntoTeachingApi.Services;
-using GetIntoTeachingApiTests.Utils;
+using GetIntoTeachingApiTests.Helpers;
 using Microsoft.Extensions.Logging;
 using Moq;
 using Xunit;
